agent-backend/requirements.txt
crewai
crewai-tools
langchain_community
python-dotenv
flask
flask-cors
litellm
mss
Pillow